+++ frontend/zac-ui/libs/features/search/src/lib/search-form/search-form.component.html
@@ -11,7 +11,7 @@
       <!-- Case (zaak) by object object. -->
       <gu-tab label="Zoeken op zaakobject">
         <div class="mt-4">
-          <gu-zaak-object-search-form (loadResult)="loadResult.emit($event)"></gu-zaak-object-search-form>
+          <gu-zaak-object-search-form (searchObjects)="searchObjects()" (selectZaakObject)="selectZaakObject($event)"></gu-zaak-object-search-form>
         </div>
       </gu-tab>
     </gu-tab-group>
