+++ frontend/zac-ui/libs/features/search/src/lib/search-form/search-form.component.ts
@@ -1,6 +1,8 @@
 import {Component, EventEmitter, Input, Output} from '@angular/core';
-import {Zaak, TableSort} from '@gu/models';
 import { PageEvent } from '@angular/material/paginator';
+import {SnackbarService} from '@gu/components';
+import {Zaak, TableSort, ZaakObject} from '@gu/models';
+import {SearchService} from '../search.service';
 
 
 /**
@@ -17,8 +19,48 @@ import { PageEvent } from '@angular/material/paginator';
   styleUrls: ['./search-form.component.scss']
 })
 export class SearchFormComponent {
+  readonly errorMessage = 'Er is een fout opgetreden bij het zoeken naar zaken.'
+
   @Input() sortData: TableSort;
   @Input() pageData: PageEvent;
   @Output() loadResult: EventEmitter<Zaak[]> = new EventEmitter<Zaak[]>();
   @Output() resultLength: EventEmitter<number> = new EventEmitter<number>();
+
+  constructor(private searchService: SearchService, private snackbarService: SnackbarService) {
+  }
+
+  //
+  // Events.
+  //
+
+  /**
+   * Gets called when zaak (case) objects are searched for.
+   */
+  searchObjects() {
+    this.loadResult.emit([]);
+  }
+
+  /**
+   * Gets called when a ZaakObject is selected.
+   * @param {ZaakObject} zaakObject
+   */
+  selectZaakObject(zaakObject: ZaakObject) {
+    const search = {
+      object: zaakObject.url
+    }
+
+    this.searchService.searchZaken(search, this.pageData?.pageIndex + 1).subscribe(
+      (data) => this.loadResult.emit(data.results as Zaak[]),
+      this.reportError.bind(this)
+    );
+  }
+
+  /**
+   * Error callback.
+   * @param {*} error
+   */
+  reportError(error: any): void {
+    this.snackbarService.openSnackBar(this.errorMessage, 'Sluiten', 'warn');
+    console.error(error);
+  }
 }
