+++ frontend/zac-ui/libs/shared/data-access/services/src/index.ts
@@ -2,3 +2,4 @@ export * from './lib/http/http-client';
 export * from './lib/shared-data-access-services.module';
 export * from './lib/user/user.service'
 export * from './lib/zaak/zaak.service'
+export * from './lib/zaak/zaak-object.service'
