+++ frontend/zac-ui/libs/shared/data-access/services/src/lib/zaak/zaak-object.service.ts
